Moved the fields Installed on, Uninstalled on and installed by from channel to movement.
