+++ LIQUIBASE/changelog/db-changelog-master.xml
@@ -28,6 +28,7 @@
 	<include file="v6.0.70\rename_inactivated_in_MT_to_active.xml" relativeToChangelogFile="true"/>
 	<include file="v6.2.3/national_endpoint_parameter.xml" relativeToChangelogFile="true"/>
 	<include file="v6.2.6/poll_payload_remove_unused_columns.xml" relativeToChangelogFile="true"/>
+	<include file="v6.2.7/move_installed_date_from_channel_to_mobileterminal.xml" relativeToChangelogFile="true"/>
 
 
 </databaseChangeLog>
