+++ model/src/main/resources/contract/MobileTerminalElement.xsd
@@ -21,6 +21,9 @@
             <xsd:element name="source" type="mobterm:MobileTerminalSource"/>
             <xsd:element name="inactive" type="xsd:boolean"/>
             <xsd:element name="archived" type="xsd:boolean"/>
+            <xsd:element name="installedOn" type="xsd:string"/>
+            <xsd:element name="uninstalledOn" type="xsd:string"/>
+            <xsd:element name="installedBy" type="xsd:string"/>
         </xsd:sequence>
     </xsd:complexType>
     
@@ -141,24 +144,6 @@
         </xsd:restriction>
     </xsd:simpleType>
     
-    <xsd:complexType name="MobileTerminalHistory">
-        <xsd:sequence>
-            <xsd:element name="events" type="mobterm:MobileTerminalEvents" minOccurs="0" maxOccurs="unbounded"/>
-            <xsd:element name="comChannels" type="mobterm:ComChannelHistory" minOccurs="0" maxOccurs="unbounded"/>
-        </xsd:sequence>
-    </xsd:complexType>
-
-    <xsd:complexType name="MobileTerminalEvents">
-        <xsd:sequence>
-            <xsd:element name="eventCode" type="mobterm:EventCode"/>
-            <xsd:element name="connectId" type="xsd:string"/>
-            <xsd:element name="serialNumber" type="xsd:string"/>
-            <xsd:element name="attributes" type="mobterm:MobileTerminalAttribute" minOccurs="0" maxOccurs="unbounded"/>
-            <xsd:element name="changeDate" type="xsd:dateTime"/>
-            <xsd:element name="comments" type="xsd:string"/>
-        </xsd:sequence>
-    </xsd:complexType>
-
     <xsd:complexType name="ComChannelHistory">
         <xsd:sequence>
             <xsd:element name="channel" type="mobterm:ComChannelHistoryAttributes" minOccurs="0" maxOccurs="unbounded"/>
