+++ rest/src/test/java/eu/europa/ec/fisheries/uvms/rest/asset/service/AssetRestResourceQueryTest.java
@@ -684,6 +684,7 @@ public class AssetRestResourceQueryTest extends AbstractAssetRestTest {
         AssetListResponse output = getWebTargetExternal()
                 .path("asset")
                 .path("list")
+                .queryParam("size", 10000)
                 .request(MediaType.APPLICATION_JSON)
                 .header(HttpHeaders.AUTHORIZATION, getTokenExternal())
                 .post(Entity.json(assetQuery), AssetListResponse.class);
